Install and configure lazygit
lazygit seems to be included in lunarvim menus by default, but also
doesn't seem to work unless lazygit is installed separately. So add it
to the nix packages and see if I use it.

Also add lazygit config because in lunarvim lazygit opens in a floating
menu that's a nuisance to get out of. Configuration means that Esc exits
the lazygit floating menu.